% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.multiple.GFS3000.R
\name{import.multiple.GFS3000}
\alias{import.multiple.GFS3000}
\title{Importing GFS3000 .csv files}
\usage{
import.multiple.GFS3000(files, names, time0s = NA)
}
\arguments{
\item{files}{list of file paths to GFS3000 .csv files}

\item{names}{list of names}

\item{time0s}{optional list of strings corresponding to desired time 0 - must be in "2021-3-18 09:00:00" format. Adds a column "TimeNo" with time in mins.}
}
\value{
A dataframe containing data from all of the GFS3000 runs
}
\description{
Made specifically for the lab324 University of Bristol GFS3000 machine which
saves csv files in a slightly odd format.
}
\details{
I've made two different functions to import files;\preformatted{1) import.single.GFS3000 - one function that assumes different runs are in the same .csv
file under different object numbers.

2) import.multiple.GFS3000 - another function that imports a list of multiple .csv files.
}
}
\examples{
#basic usage of import.multiple.GFS3000

DF1 = import.multiple.GFS3000(c("file1.csv", "file2.csv"), name=c("Col-0","Col-0"), time0s=c("2019-11-19 09:00:00","2019-11-19 09:00:00"))
}
