% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.single.GFS3000.R
\name{import.single.GFS3000}
\alias{import.single.GFS3000}
\title{Importing GFS3000 .csv files}
\usage{
import.single.GFS3000(file, names = NA, time0s = NA)
}
\arguments{
\item{file}{path to GFS3000 .csv file}

\item{names}{optional list of strings corresponding to object numbers}

\item{time0s}{optional list of strings corresponding to desired time 0 - must be in "2021-3-18 09:00:00" format. Adds a column "TimeNo" with time in mins.}
}
\value{
A dataframe containing data from the GFS3000 run
}
\description{
Made specifically for the lab324 University of Bristol GFS3000 machine which
saves csv files in a slightly odd format.
}
\details{
I've made two different functions to import files;\preformatted{1) import.single.GFS3000 - one function that assumes different runs are in the same .csv
file under different object numbers.

2) import.multiple.GFS3000 - another function that imports a list of multiple .csv files.
}
}
\examples{
#basic usage of import.single.GFS3000

DF1 = import.single.GFS3000("file1.csv", names=c("Col-0"), time0s=c("2019-11-19 09:00:00"))

#usage of import.single.GFS3000 to read a file with multiple objects
#Here object 0001 corresponds to Col-0
#and object 0002 corresponds to q1124

DF1 = import.single.GFS3000("file1.csv", names=c("Col-0","q1124"), time0s=c("2019-11-19 09:00:00","2019-11-19 12:00:00"))
}
